Fix the Octree Quantization Algorithm and refactor Node class
