+++ RasterPaint/RasterPaint/RasterPaint.csproj
@@ -145,8 +145,10 @@
       <SubType>Code</SubType>
     </Compile>
     <Compile Include="Converters\MultiValueConverter.cs" />
+    <Compile Include="Objects\Node.cs" />
     <Compile Include="Objects\Octree.cs" />
     <Compile Include="Utilities\ColorReduction.cs" />
+    <Compile Include="Utilities\IdHelper.cs" />
     <Compile Include="Views\ColorReductionWindow.xaml.cs">
       <DependentUpon>ColorReductionWindow.xaml</DependentUpon>
     </Compile>
