+++ RasterPaint/RasterPaintTests/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="WriteableBitmapEx" version="1.5.0.0" targetFramework="net452" />
+</packages>
